requirements.txt
# Core Flask dependencies
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-JWT-Extended>=4.5.0
Flask-CORS>=4.0.0

# Database
psycopg2-binary>=2.9.0

# External API integrations
requests>=2.31.0
aiohttp>=3.8.0

# Background tasks
celery[redis]>=5.3.0
redis>=4.6.0

# Environment and configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-flask>=1.2.0
pytest-cov>=4.1.0

# Development
flake8>=6.0.0
black>=23.0.0